feat: 拆分 cmd.go 中的子命令到不同的文件
