+++ .gitignore
@@ -24,9 +24,8 @@ cmd/vmr/vmr
 cmd/vmr/vmr.exe
 cmd/vmr/vm
 cmd/vmr/vm.exe
-manager
-vmr
-vm
+
+bin/
 
 build_win.sh
 .DS_Store
