+++ cmd/vmr/main.go
@@ -22,7 +22,7 @@
 package main
 
 import (
-	"github.com/gvcgo/version-manager/cmd"
+	"github.com/gvcgo/version-manager/cmd/vmr/internal/cli"
 	_ "github.com/gvcgo/version-manager/pkgs/conf"
 )
 
@@ -36,6 +36,6 @@ To be compiled.
 */
 func main() {
 	// os.Setenv(conf.VMReverseProxyEnvName, "https://gvc.1710717.xyz/proxy/")
-	cli := cmd.New(GitTag, GitHash)
-	cli.Run()
+	cmd := cli.New(GitTag, GitHash)
+	cmd.Run()
 }
